Move the model registry to a separate file
