+++ src/core/ts/web-worker-scripts/translation-worker.js/getBergamotModelsForLanguagePair.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { ModelRegistry } from "../../config";
+import { ModelRegistry } from "../../modelRegistry";
 import { ModelDownloadProgress } from "../../background-scripts/background.js/lib/BergamotTranslatorAPI";
 import {
   DownloadedFile,
